LayoutTests/platform/ios-simulator/editing/pasteboard/copy-paste-bidi-expected.txt

These tests to see if bidi text with a base writing direction of right to left is put onto the paste board in logical (DOM) order.

first test - before:
| "1234שדגכ:"

first test - after:
| <span>
|   style="-webkit-text-size-adjust: auto; background-color: rgba(255, 255, 255, 0);"
|   "1234שדגכ:<#selection-caret>"

second test - before:
| "
"
| <div>
|   "1234שדגכ:"
| "
"
| <div>
|   style="direction: rtl;"
|   "1234שדגכ:"
| "
"

second test - after:
| "
"
| <div>
|   <div>
|     <span>
|       style="-webkit-text-size-adjust: auto; background-color: rgba(255, 255, 255, 0);"
|       "1234שדגכ:"
|   <div>
|     style="direction: rtl;"
|     <span>
|       style="-webkit-text-size-adjust: auto; background-color: rgba(255, 255, 255, 0);"
|       "1234שדגכ:<#selection-caret>"
| "
"
